Removed/Relocated unnecessary Test Files
