+++ Backend/src/test/java/com/cisco/TrendSight/service/JwtSecretMakerTest.java
@@ -1,4 +1,4 @@
-package com.cisco.TrendSight;
+package com.cisco.TrendSight.service;
 
 
 import io.jsonwebtoken.Jwts;
